+++ hydra-node/test/Hydra/Model/MockChain.hs
@@ -338,9 +338,7 @@ fixedTimeHandleIndefiniteHorizon = do
 scriptLedger ::
   Ledger Tx
 scriptLedger =
-  -- NOTE: This mock only validates scripts, so re-application reuses the same
-  -- logic.
-  Ledger{applyTransactions, reapplyTransactions = applyTransactions}
+  Ledger{applyTransactions}
  where
   -- XXX: We could easily add 'slot' validation here and this would already
   -- emulate the dropping of outdated transactions from the cardano-node